GetWithdrawals + tests

+++ internal/models/account.go
@@ -12,7 +12,7 @@ type Account struct {
 
 type Withdrawal struct {
 	OrderNum    OrderNum  `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
